refactor Clauses and Automas to work based on continuous probabilities (#10)

+++ clause.go
@@ -1,17 +1,20 @@
 package getlin
 
 type Clause interface {
-	// Output produces the Clause's binary label prediction given a binary
-	// feature Vector.
-	Output([]uint8) uint8
-	// States returns the internal state pointers of this Clause's TAs. The
-	// first list returned represents states of negative polarity and the second
-	// list returned represents states of positive polarity.
-	States() ([]int, []int)
-	// TypOne applies Type 1 Feedback to the Clause, given a binary feature
-	// Vector and its own label prediction.
-	TypOne([]uint8, uint8)
-	// TypTwo applies Type 2 Feedback to the Clause, given a binary feature
-	// Vector and its own label prediction.
-	TypTwo([]uint8, uint8)
+	// Output produces the Clause's weight prediction given some probabilistic
+	// feature Vector. Features and prediction are within the range [0.0, 1.0].
+	Output([]float32) float32
+	// States returns the current state pointers of this Clause's Automas within
+	// the range [0.0, 1.0]. The first list returned represents states of
+	// negative polarity and the second list returned represents states of
+	// positive polarity.
+	States() ([]float32, []float32)
+	// TypOne applies Type 1 Feedback to the Clause, given some probabilistic
+	// feature Vector and its own weight prediction. Features and prediction are
+	// within the range [0.0, 1.0].
+	TypOne([]float32, float32)
+	// TypTwo applies Type 2 Feedback to the Clause, given some probabilistic
+	// feature Vector and its own weight prediction. Features and prediction are
+	// within the range [0.0, 1.0].
+	TypTwo([]float32, float32)
 }
